+++ Sources/ShapeUp/Corner/RelativeCorner+Animatable.swift
@@ -11,11 +11,39 @@ extension RelativeCorner: AnimatableByProperty {
     public static var animatableProperties: some AnimatableProperty<Self> {
         \.anchor
         \.offset
+        \.style.radius
+        \.style.concaveInset
+        \.style.relativeCorners.anchors
+        \.style.relativeCorners.offsets
+        \.style.relativeCorners.cornerStyles.radii
+        \.style.relativeCorners.cornerStyles.concaveInsets
+    }
+}
+
+extension Array where Element == RelativeCorner {
+    var anchors: [RectAnchor] {
+        get {
+            map(\.anchor)
+        }
+        set {
+            self = self.update(with: newValue) { element, newValue in
+                var updatedElement = element
+                updatedElement.anchor = newValue
+                return updatedElement
+            }
+        }
     }
     
-    public static var recursiveAnimatableProperties: some AnimatableProperty<Self> {
-        \.style
-        \.style.relativeCorners
-        \.style.relativeCorners.cornerStyles
+    var offsets: [Vector2] {
+        get {
+            map(\.offset)
+        }
+        set {
+            self = self.update(with: newValue) { element, newValue in
+                var updatedElement = element
+                updatedElement.offset = newValue
+                return updatedElement
+            }
+        }
     }
 }